CardioAI_Groq/backend/requirements.txt
# CardioAI — Python requirements
# Install: pip install -r requirements.txt
# No version pins — works with Python 3.10, 3.11, 3.12, 3.13

fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv

scikit-learn
pandas
numpy
joblib

requests

sqlalchemy
pymysql
cryptography
